extract function into its own file
