+++ packages/mui-system/src/Unstable_Grid/index.ts
@@ -3,4 +3,4 @@ export { default as createGrid } from './createGrid';
 export * from './GridProps';
 export { default as gridClasses } from './gridClasses';
 export * from './gridClasses';
-export { traverseBreakpoints as unstable_traverseBreakpoints } from './gridGenerator';
+export { traverseBreakpoints as unstable_traverseBreakpoints } from './traverseBreakpoints';
